refactor: Move request verb annotations to data.http.requestverb
